+++ 411-resources-main/HW/HW3 Design/wildlife_tracker/habitat_management/habitat.py
@@ -1,4 +1,6 @@
 from typing import Any, List, Optional
+from animal_management.animal import Animal
+
 
 class Habitat:
