+++ 411-resources-main/HW/HW3 Design/wildlife_tracker/migration_tracking/migration.py
@@ -1,5 +1,16 @@
-from typing import Any
+from typing import Any, List, Optional
+from migration_path import MigrationPath
+
 
 class Migration:
 
-    pass
+    def __init__(self, 
+                migration_id: int,
+                migration_path: MigrationPath) -> None:
+        pass
+
+    def get_migration_details(migration_id: int) -> dict[str, Any]:
+        pass
+
+    def update_migration_details(migration_id: int, **kwargs: Any) -> None:
+        pass
