+++ 411-resources-main/HW/HW3 Design/wildlife_tracker/migration_tracking/migration_manager.py
@@ -1,6 +1,58 @@
-from typing import Optional
+from typing import Any, List, Optional
+from typing import Any, List, Optional
+from habitat_management.habitat import Habitat
+from migration_tracking.migration import Migration
+from migration_path import MigrationPath
 
 
 class MigrationManager:
+    def __init__(self) -> None:
+        migrations: dict[int, Migration] = {}
+        paths: dict[int, MigrationPath] = {}
 
+
+def cancel_migration(migration_id: int) -> None:
     pass
+
+def create_migration_path(species: str, start_location: Habitat, destination: Habitat, duration: Optional[int] = None) -> None:
+    pass
+
+def get_migration_by_id(migration_id: int) -> Migration:
+    pass
+
+def get_migration_path_by_id(path_id: int) -> MigrationPath:
+    pass
+
+def get_migration_paths() -> list[MigrationPath]:
+    pass
+
+def get_migration_paths_by_destination(destination: Habitat) -> list[MigrationPath]:
+    pass
+
+def get_migration_paths_by_species(species: str) -> list[MigrationPath]:
+    pass
+
+def get_migration_paths_by_start_location(start_location: Habitat) -> list[MigrationPath]:
+    pass
+
+def get_migrations() -> list[Migration]:
+    pass
+
+def get_migrations_by_current_location(current_location: str) -> list[Migration]:
+    pass
+
+def get_migrations_by_migration_path(migration_path_id: int) -> list[Migration]:
+    pass
+
+def get_migrations_by_start_date(start_date: str) -> list[Migration]:
+    pass
+
+def get_migrations_by_status(status: str) -> list[Migration]:
+    pass
+
+def remove_migration_path(path_id: int) -> None:
+    pass
+
+def schedule_migration(migration_path: MigrationPath) -> None:
+    pass
+
