+++ 411-resources-main/HW/HW3 Design/wildlife_tracker/migration_tracking/migration_path.py
@@ -1,5 +1,20 @@
-from typing import Optional
+from typing import Any, List, Optional
+from habitat_management.habitat import Habitat
 
 class MigrationPath:
 
-    pass
+    def __init__(self, 
+                 path_id: int,
+                 size: int,
+                 species: str,
+                 start_date: str,
+                 start_location: Habitat,
+                 status: str = "Scheduled") -> None:
+        pass
+
+
+    def get_migration_path_details(path_id) -> dict:
+        pass
+
+    def update_migration_path_details(path_id: int, **kwargs) -> None:
+        pass
